Check env val in main

+++ pritunl_client/__main__.py
@@ -246,6 +246,9 @@ def get_env():
     with open(env_path, 'r') as env_file:
         env = json.loads(env_file.read())
 
+    if not env.get('VPN_CONF'):
+        raise ValueError('Invalid environment file')
+
     os.remove(env_path)
     return env
